provider/data: Split batch from data
